[#219] Refactor: MemberFacade에 updateProfileImage 추가 및 S3 로직 분리 완료
