+++ src/main/java/soon/fridgely/domain/member/service/MemberService.java
@@ -1,14 +1,12 @@
 package soon.fridgely.domain.member.service;
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
-import org.springframework.web.multipart.MultipartFile;
 import soon.fridgely.domain.EntityStatus;
 import soon.fridgely.domain.member.dto.command.MemberInfo;
 import soon.fridgely.domain.member.dto.response.MemberProfileResponse;
@@ -17,11 +15,8 @@ import soon.fridgely.domain.member.entity.MemberRole;
 import soon.fridgely.domain.member.repository.MemberRepository;
 import soon.fridgely.global.support.exception.CoreException;
 import soon.fridgely.global.support.exception.ErrorType;
-import soon.fridgely.global.support.image.ImageManager;
 import soon.fridgely.global.support.image.event.ImageDeleteEvent;
-import soon.fridgely.global.support.logging.SlackMarkers;
 
-@Slf4j
 @RequiredArgsConstructor
 @Service
 public class MemberService {
@@ -29,7 +24,6 @@ public class MemberService {
     private final MemberRepository memberRepository;
     private final PasswordEncoder passwordEncoder;
     private final ApplicationEventPublisher eventPublisher;
-    private final ImageManager imageManager;  // Task 8에서 제거
 
     @Transactional
     public Member register(MemberInfo memberInfo) {
@@ -65,30 +59,6 @@ public class MemberService {
         member.updateProfileImage(newImageUrl);
         if (StringUtils.hasText(oldImageUrl) && !oldImageUrl.equals(newImageUrl)) {
             eventPublisher.publishEvent(new ImageDeleteEvent(oldImageUrl));
-        }
-    }
-
-    // Task 8에서 MemberFacade로 이동 후 제거
-    public void updateProfileImage(long memberId, MultipartFile file) {
-        if (file == null || file.isEmpty()) {
-            throw new CoreException(ErrorType.INVALID_REQUEST);
-        }
-        String uploadedUrl = imageManager.upload(file);
-        try {
-            updateProfileImage(memberId, uploadedUrl);
-        } catch (Exception e) {
-            rollbackImageUpload(uploadedUrl);
-            throw e;
-        }
-    }
-
-    private void rollbackImageUpload(String imageUrl) {
-        if (imageUrl != null) {
-            try {
-                imageManager.delete(imageUrl);
-            } catch (Exception e) {
-                log.warn(SlackMarkers.SYSTEM, "[Member] 이미지 롤백 실패 - 수동 정리 필요 (ImageUrl={})", imageUrl, e);
-            }
         }
     }
 }
